app/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
werkzeug==2.3.0
python-dotenv==1.0.0
stripe==5.4.0
twilio==8.2.0
firebase-admin==6.1.0
Pillow==10.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.1
google-auth==2.20.0
sqlalchemy==2.0.15